yess-backend/requirements.txt
# Основные зависимости
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
gunicorn>=23.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.9

# Безопасность и токены
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.0
PyJWT>=2.9.0

# Валидация и конфигурация
pydantic>=2.9.0
pydantic[dotenv]>=2.9.0
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
email-validator>=2.2.0

# Асинхронные драйверы
asyncpg>=0.29.0

# Redis / Celery
redis>=5.0.0
celery>=5.4.0

# Работа с формами и файлами
python-multipart>=0.0.12
aiofiles>=24.1.0
qrcode[pil]>=8.0
Pillow>=10.4.0

# Ограничение запросов
slowapi>=0.1.9

# Геолокация
geopy>=2.4.1

# HTTP клиент
requests>=2.32.0
urllib3>=2.2.0

# Мониторинг
prometheus-client>=0.21.0
sentry-sdk>=2.19.0

# Тесты
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
httpx>=0.27.0
locust>=2.28.0

# Инструменты
coverage>=7.6.0
responses>=0.25.0

# Линтеры
flake8>=7.1.0
black>=24.8.0
mypy>=1.11.0

# Профилирование
py-spy>=0.3.14
memory-profiler>=0.61.0

# Генераторы данных
faker>=30.0.0
factory-boy>=3.3.0

# Безопасность анализ
safety>=3.3.0
bandit>=1.7.8
